+++ libfreedv/ofdm_internal.h
@@ -43,8 +43,8 @@
 #define TAU         (2.0f * M_PI)
 #define ROT45       (M_PI / 4.0f)
 
-#define cmplx(value) (std::complex<float>{cos(value), sin(value)})
-#define cmplxconj(value) (std::complex<float>{cos(value), -sin(value)})
+#define cmplx(value) (std::complex<float>{cosf(value), sinf(value)})
+#define cmplxconj(value) (std::complex<float>{cosf(value), -sinf(value)})
 
 namespace FreeDV
 {
